+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Changelog](http://keepachangelog.com/en/1.0.0/) and this project
 adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - xxxx-xx-xx
+### Changed
+- Default value for baseURL
 
 ## [v1.1.0] - 2018-11-07
 ### Added